requirements.txt
mcp>=1.2.0
google-generativeai>=0.8.0
httpx>=0.25.0
python-dotenv>=1.0.0